test(oracles): split baseline compatibility regressions
Why:
- keep CoinTracking baseline regression coverage under the repo's test size boundary
- isolate export-shape compatibility cases from the core baseline row parser tests

What:
- move real export compatibility regressions into a dedicated baseline export compatibility test module
- leave the existing baseline row parser tests focused on shared parsing invariants

Checks:
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run pytest -q --no-cov tests/oracles/cointracking/test_baseline_rows.py tests/oracles/cointracking/test_baseline_export_compatibility.py tests/oracles/workflows/test_verification_service.py
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run ruff check tests/oracles/cointracking/test_baseline_rows.py tests/oracles/cointracking/test_baseline_export_compatibility.py tests/oracles/workflows/test_verification_service.py
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run pylint --rcfile=.pylintrc-tests tests/oracles/cointracking/test_baseline_rows.py tests/oracles/cointracking/test_baseline_export_compatibility.py tests/oracles/workflows/test_verification_service.py
